backend/requirements.txt
asgiref==3.5.2
autopep8==1.7.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
Django==2.2.16
django-debug-toolbar==3.2
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-base64==2.0
drf-extra-fields==3.4.1
gunicorn==20.1.0
idna==3.4
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
npm==0.1.1
oauthlib==3.2.2
optional-django==0.1.0
Pillow==9.2.0
psycopg2-binary==2.8.6
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.5.0
pyparsing==3.0.9
python3-openid==3.2.0
pytz==2022.4
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
toml==0.10.2
typing-extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
zipp==3.9.0
